requirements.txt
# requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
jsonschema>=4.21.0
tenacity>=8.2.3
